--- apps/web/modules/ee/teams/team-list/components/EditMemberships/index.ts
@@ -1,2 +1,0 @@
-// export { EditMembershipsClient } from "./EditMembershipscClient";
-export { EditMemberships } from "./EditMemberships";
