+++ apps/web/modules/ee/teams/team-list/components/edit-memberships/index.ts
@@ -1,0 +1,1 @@
+export { EditMemberships } from "./edit-memberships";
