+++ apps/web/modules/ee/teams/team-list/components/invite-member/ShareInviteModal.tsx
